Resolved issue #500 ('Matchmaking server's battle statistics link gives error 404'), also uptime counter overflow is now fixed
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@1040 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/server.h
@@ -29,9 +29,9 @@
 #define PACKET_SIZE_LIMIT 65536
 #endif
 
-inline long get_time_diff(const NLtime &x, const NLtime &y)
+inline double get_time_diff(const NLtime &x, const NLtime &y)
 {
-    return (y.seconds - x.seconds) * 1000 + (y.mseconds - x.mseconds);
+    return (double)(y.seconds - x.seconds) * 1000.0 + (double)(y.mseconds - x.mseconds);
 }
 
 class ServerDispatch;
